Split functions in cli.go
